dialog cleanup
